Interface Orientation update, Item creation update
Interface Orientation update for iPad and iPhone. Image view and camera
button disabled in landscape mode for iPhone. The image picker is
presented in a popover for the iPad.

Instead of creating an item randomly, Homepwner is updated to present
the BNRDetailViewController modally when the user creates a new BNRItem.